build: Install module_common_types.h and dependencies
This is needed for at least the AudioFrame class.

Unfortunately, this does add a bit of ugliness because
module_common_types.h has video bits that are hidden behind our own
define, which now becomes part of pkg-config CFLAGS.

This could be made less ugly, potentially, but I'm not sure how right
now.